// src/lib/tamtree/v1.d.ts
export interface components {
    schemas: {
// ...
        FlowPage: {
            /** Items */
            items: components["schemas"]["FlowOut"][];
            /** Next Cursor */
            next_cursor: string | null;
        };
// ...
        RunAssetPage: {
            /** Items */
            items: components["schemas"]["RunAssetOut"][];
            /** Next Cursor */
            next_cursor: string | null;
        };
// ...
        RunEventV1Page: {
            /** Items */
            items: components["schemas"]["RunEventOut"][];
            /** Next Cursor */
            next_cursor: string | null;
        };
// ...
        RunFlowIn: {
            /** End User Ref */
            end_user_ref?: string | null;
            /** Input */
            input?: {
                [key: string]: unknown;
            } | null;
            /** Metadata */
            metadata?: {
                [key: string]: string;
            } | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}